lambda/requirements.txt
Pillow==10.4.0
piexif==1.1.3
requests==2.32.3
boto3==1.34.162